Major updates: 1) Refactor forum management: add forumAdmin to replace forumManagement page. 2) Support sorting of forum and forum groups. 3) Support sorting of tags and chat channels

+++ src/main/java/com/github/chipolaris/bootforum/jsf/bean/TagManagement.java
@@ -1,11 +1,13 @@
 package com.github.chipolaris.bootforum.jsf.bean;
 
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
+import javax.faces.component.UIComponent;
+import javax.faces.context.FacesContext;
+import javax.faces.convert.Converter;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,7 +21,13 @@ import com.github.chipolaris.bootforum.jsf.util.JSFUtils;
 import com.github.chipolaris.bootforum.service.AckCodeType;
 import com.github.chipolaris.bootforum.service.GenericService;
 import com.github.chipolaris.bootforum.service.ServiceResponse;
+import com.github.chipolaris.bootforum.service.TagService;
 
+/**
+ * 
+ * Main backing bean for /admin/tagManagement.xhtml page
+ *
+ */
 @Component
 @Scope("view")
 public class TagManagement {
@@ -28,6 +36,9 @@ public class TagManagement {
 	
 	@Resource
 	private GenericService genericService;
+	
+	@Resource
+	private TagService tagService;
 	
 	@Resource 
 	private CacheManager cacheManager;
@@ -59,7 +70,7 @@ public class TagManagement {
 	
 	@PostConstruct
 	private void postConstruct() {
-		this.tags = genericService.getAllEntities(Tag.class).getDataObject();
+		this.tags = genericService.getEntities(Tag.class, Collections.emptyMap(), "sortOrder", false).getDataObject();
 		this.setNewTag(new Tag());
 	}
 	
@@ -68,15 +79,14 @@ public class TagManagement {
 		logger.info("Creating tag " + newTag.getLabel());
 		
 		// check if tag.label exists
-		Map<String, Object> filters = new HashMap<>();
-		filters.put("label", newTag.getLabel());
-		Long tagCount = genericService.countEntities(Tag.class, filters).getDataObject();
+		Long tagCount = genericService.countEntities(Tag.class, 
+				Collections.singletonMap("label", newTag.getLabel())).getDataObject();
 		if(tagCount > 0) {
 			JSFUtils.addErrorStringMessage(null, String.format("Tag's label '%s' already exists", newTag.getLabel()));
 			return;
 		}
 		
-    	ServiceResponse<Long> response = genericService.saveEntity(newTag);
+		ServiceResponse<Long> response = tagService.createNewTag(newTag);
     	
     	if(response.getAckCode().equals(AckCodeType.SUCCESS)) {
     		JSFUtils.addInfoStringMessage(null, String.format("Tag %s created", newTag.getLabel()));
@@ -122,6 +132,13 @@ public class TagManagement {
 		logger.info("Deleting tag " + selectedTag.getLabel());
 		
     	if(this.selectedTag != null) {
+    		
+    		if(tagService.countDiscussionsForTag(selectedTag).getDataObject() > 0) {
+    			JSFUtils.addErrorStringMessage(null, 
+    					String.format("Unable to delete Tag '%s'. It is being referenced by one or more discussions", this.selectedTag.getLabel()));
+    			return;
+    		}
+    		
 	    	// 
 	    	ServiceResponse<Void> response = genericService.deleteEntity(this.selectedTag);
 	    	
@@ -141,5 +158,54 @@ public class TagManagement {
     	else {
     		JSFUtils.addErrorStringMessage(null, "deleteTag() called on null selectedTag");
     	}
+	}
+	
+	public void tagOrderSubmit() {
+		
+		logger.info("Ordering tags");
+		
+		for(int i = 0; i < tags.size(); i++) {
+			Tag tag = tags.get(i);
+			tag.setSortOrder(i + 1);
+			genericService.updateEntity(tag);
+		}
+		
+		JSFUtils.addInfoStringMessage(null, "Tags (re)ordered");
+	}
+	
+	/**
+	 * Tag converter
+	 */
+	private Converter<Tag> tagConverter = new Converter<Tag>() {
+		
+		@Override
+		public Tag getAsObject(FacesContext context, UIComponent component, String idStr) {
+			Long id;
+			try {
+				id = new Long(idStr);
+			} 
+			catch (NumberFormatException e) {
+				return null;
+			}
+			
+			// traverse through the collection of tags
+			// and find the object that have the given id
+			for(Tag tag : tags) {
+				if(tag.getId().equals(id)) {
+					return tag;
+				}
+			}
+
+			return null;
+		}
+
+		@Override
+		public String getAsString(FacesContext context, UIComponent component, Tag value) {
+			return value.getId().toString();
+		}
+	};
+	
+	public Converter<Tag> getTagConverter() {
+		return tagConverter;
 	}
 }
